后台不能删除默认客户组

+++ beike/Admin/Http/Controllers/CustomerController.php
@@ -79,6 +79,9 @@ class CustomerController extends Controller
 
     public function destroy(Request $request, int $customerId)
     {
+        if ($customerId == setting('base.default_customer_group_id')) {
+            return json_fail(trans('customer_group.default_cannot_delete'));
+        }
         CustomerRepo::delete($customerId);
 
         return json_success(trans('common.deleted_success'));
